+++ src/links.rs
@@ -62,4 +62,5 @@ pub mod direct {
     pub const SHOW: &'static str = "https://api.twitter.com/1.1/direct_messages/show.json";
     pub const RECEIVED: &'static str = "https://api.twitter.com/1.1/direct_messages.json";
     pub const SENT: &'static str = "https://api.twitter.com/1.1/direct_messages/sent.json";
+    pub const SEND: &'static str = "https://api.twitter.com/1.1/direct_messages/new.json";
 }